What happens when a block is pushed, whether from peer or by own generator:
- unconfirmed txes effected, unconfirmed txes and balance reflected in ui.
- generator adds official txes to the block.
- pushBlock
-- earlier effected unconfirmed txes rolled back.
-- actual block transactions effected.
-- block is now broadcasted.

Changes:
On each 2nd block of the day, when giving interest -
The generator does a preliminary check on what the balances would be if interest were negative, then goes through the transactions to see which will bring the balance to negative, and rejects those transactions. In this way, the official block transactions will not bring any account balances to negative.
Interest is then given after the rollback of the earlier unconfirmed txes, before the effecting of the official block transactions.